Add option to save description as HTML.
Some authors upload images for stories, and use the description
field for the actual text. If this is done, having just the plain
text version of the story can omit formatting.

This adds a -H flag - when used, the downloader will capture both
the plain and the HTML versions of the submission description.